# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "servers": [
    {
      "url": "http://localhost:8000/"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "description": "",
        "responses": {
          "504": {
            "description": "Gateway Timeout"
          }
        }
      }
    },
    "/msg": {
      "post": {
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "Author": {
                    "example": "any"
                  },
                  "Title": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/del": {
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/patch": {
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "example": "any"
                  },
                  "Author": {
                    "example": "any"
                  },
                  "Title": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/books": {
      "get": {
        "summary": "Get all books.",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/books/{id}": {
      "get": {
        "summary": "Find a book with a specified ID.",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "summary": "Create a book. Returns the Entry and the ID of the object.",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "Author": {
                    "example": "any"
                  },
                  "Title": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update a book with a specified ID.",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "Author": {
                    "example": "any"
                  },
                  "Title": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a book with an ID.",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "openapi": {
        "type": "string",
        "example": "3.0.0"
      },
      "info": {
        "type": "object",
        "properties": {
          "version": {
            "type": "string",
            "example": "0.0.1"
          },
          "title": {
            "type": "string",
            "example": "My Little API"
          },
          "description": {
            "type": "string",
            "example": "RESTful is magic"
          }
        }
      },
      "servers": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "url": {
              "type": "string",
              "example": ""
            },
            "description": {
              "type": "string",
              "example": "gentoo computer"
            }
          }
        }
      },
      "components": {
        "type": "object",
        "properties": {
          "schemas": {
            "type": "object",
            "properties": {
              "Book": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "object"
                  },
                  "required": {
                    "type": "array",
                    "example": [
                      "Author",
                      "Title"
                    ],
                    "items": {
                      "type": "string"
                    }
                  },
                  "properties": {
                    "type": "object",
                    "properties": {
                      "_id": {
                        "type": "object",
                        "properties": {
                          "type": {
                            "type": "string",
                            "example": "string"
                          },
                          "description": {
                            "type": "string",
                            "example": "MongoDB document ID"
                          }
                        }
                      },
                      "Author": {
                        "type": "object",
                        "properties": {
                          "type": {
                            "type": "string",
                            "example": "string"
                          },
                          "description": {
                            "type": "string",
                            "example": "Author of the book"
                          }
                        }
                      },
                      "Title": {
                        "type": "object",
                        "properties": {
                          "type": {
                            "type": "string",
                            "example": "string"
                          },
                          "description": {
                            "type": "string",
                            "example": "Book title"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}